Move observations into activities
